[lldb] Update Swift plugins for ValueObject move
