+++ lldb/source/Plugins/ExpressionParser/Swift/SwiftExpressionVariable.cpp
@@ -13,12 +13,12 @@
 #include "SwiftExpressionVariable.h"
 
 #include "lldb/Core/Value.h"
-#include "lldb/Core/ValueObjectConstResult.h"
 #include "lldb/Target/ExecutionContext.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Utility/ConstString.h"
 #include "lldb/Utility/DataExtractor.h"
 #include "lldb/Utility/Stream.h"
+#include "lldb/ValueObject/ValueObjectConstResult.h"
 #include "clang/AST/ASTContext.h"
 
 using namespace lldb_private;
