+++ lldb/source/Plugins/Language/Swift/SwiftArray.cpp
@@ -15,10 +15,10 @@
 #include "Plugins/LanguageRuntime/Swift/SwiftLanguageRuntime.h"
 #include "Plugins/TypeSystem/Clang/TypeSystemClang.h"
 #include "Plugins/TypeSystem/Swift/TypeSystemSwiftTypeRef.h"
-#include "lldb/Core/ValueObjectConstResult.h"
 #include "lldb/DataFormatters/FormattersHelpers.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Target/Target.h"
+#include "lldb/ValueObject/ValueObjectConstResult.h"
 
 // FIXME: we should not need this
 #include "Plugins/Language/ObjC/Cocoa.h"
