+++ lldb/source/Plugins/Language/Swift/SwiftOptionSet.cpp
@@ -14,11 +14,11 @@
 
 #include "Plugins/TypeSystem/Clang/TypeSystemClang.h"
 #include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
-#include "lldb/Core/ValueObject.h"
 #include "lldb/Symbol/CompilerType.h"
 #include "lldb/Utility/LLDBLog.h"
 #include "lldb/Utility/Log.h"
 #include "lldb/Utility/StreamString.h"
+#include "lldb/ValueObject/ValueObject.h"
 
 #include "clang/AST/Decl.h"
 #include "llvm/ADT/StringRef.h"
