+++ lldb/source/Plugins/Language/Swift/SwiftUnsafeTypes.h
@@ -13,8 +13,8 @@
 #ifndef liblldb_SwiftUnsafeTypes_h_
 #define liblldb_SwiftUnsafeTypes_h_
 
-#include "lldb/Core/ValueObject.h"
 #include "lldb/Utility/Stream.h"
+#include "lldb/ValueObject/ValueObject.h"
 
 namespace lldb_private {
 namespace formatters {
